requirements.txt
django
djangorestframework
pygments

